+++ src/components/organisms/AddProduct/addProduct.module.css
@@ -2,7 +2,7 @@
   padding-top: 30px;
 }
 
-.wrapper-form > h2 {
+.wrapper-form > h3 {
   font-size: 1.2rem;
   line-height: 1.4rem;
   color: var(--darkgray-color);
